+++ src/main/java/team12/stockist/controller/UserController.java
@@ -16,6 +16,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import team12.stockist.service.UsageRecordService;
 import team12.stockist.service.UserService;
 import team12.stockist.validator.UserValidator;
 
@@ -23,6 +26,8 @@ import team12.stockist.validator.UserValidator;
 @Controller
 public class UserController {
 
+	@Autowired
+	UsageRecordService usageRecordService;
 	@Autowired
 	UserService userService;
 	@Autowired
@@ -49,16 +54,29 @@ public class UserController {
 	}
 
 	@RequestMapping(value = "/create", method = RequestMethod.POST)
-	public ModelAndView createUserPage(@ModelAttribute @Valid User user, BindingResult result) {
+	public ModelAndView createUserPage(@ModelAttribute @Valid User user, BindingResult result, final RedirectAttributes redirectAttribute) {
 
-		if (result.hasErrors())
+		/*if (result.hasErrors())
 			return new ModelAndView("user-new");
+		 */
 		ModelAndView mav = new ModelAndView();
-
-		userService.createUser(user);
-		mav.setViewName("redirect:/admin/user/list");
-
-		return mav;
+		
+		if(result.hasErrors()) {
+			mav.setViewName("redirect:/admin/user/create");
+			redirectAttribute.addFlashAttribute("useralreadyexists", "All fields must be completed");
+			return mav;
+		}
+		
+		if(userService.userAlreadyExists(user)) {
+			mav.setViewName("redirect:/admin/user/create");
+			redirectAttribute.addFlashAttribute("useralreadyexists", "User already exists");
+			return mav;
+		}
+		else {
+			userService.createUser(user);
+			mav.setViewName("redirect:/admin/user/list");
+			return mav;
+		}
 	}
 
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
@@ -72,14 +90,26 @@ public class UserController {
 
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.POST)
 	public ModelAndView editUserPage(@ModelAttribute @Valid User user, BindingResult result, @PathVariable String id) {
-		if (result.hasErrors())
-			return new ModelAndView("user-edit");
+		
+		ModelAndView mav = new ModelAndView();
+		
+		if(result.hasErrors()) {
+			mav.setViewName("user-edit");
+			mav.addObject("useralreadyexists", "All fields must be completed");
+			return mav;
+		}
+		if(userService.userAlreadyExists(user)) {
+			mav.setViewName("user-edit");
+			mav.addObject("useralreadyexists", "User already exists");
+			return mav;
+		}
+		else {
+			userService.updateUser(user);
+			mav.setViewName("redirect:/admin/user/list");
+			return mav;
+		}
+	
 
-		ModelAndView mav = new ModelAndView("redirect:/admin/user/list");
-
-		userService.updateUser(user);
-
-		return mav;
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
@@ -87,6 +117,12 @@ public class UserController {
 		ModelAndView mav = new ModelAndView("redirect:/admin/user/list");
 		int iid = Integer.parseInt(id);
 		User user = userService.findUserById(iid);
+
+		if (usageRecordService.usageRecordisNotDeletable(user)) {
+			mav.addObject("userdeleteerror", "Cannot delete user. Please update user role to inactive to delete.");
+			return mav;
+		}
+
 		userService.deleteUser(user);
 		return mav;
 	}
